Add: new functionality (add testplan, add testplan entry)
