+++ pytest_testrail/__init__.py
@@ -1,1 +1,3 @@
 # -*- coding: UTF-8 -*-
+from pytest_testrail.functions import pytestrail, testrail
+
